Crate gluon

This crate contains contains the implementation for the gluon programming language.

Gluon is a programming language suitable for embedding in an existing application to extend its behaviour. For information about how to use this library the best resource currently is the tutorial which contains examples on how to write gluon programs as well as how to run them using this library.

Reexports

pub extern crate gluon_vm as vm;

pub extern crate gluon_base as base;

pub extern crate gluon_parser as parser;

pub extern crate gluon_check as check;

Modules

compiler_pipeline
Advanced compiler pipeline which ensures that the compilation phases are run in order even if not the entire compilation procedure is needed.

import
Implementation of the import macro.

io

Structs

Compiler
Type which makes parsing, typechecking and compiling an AST into bytecode

RootedThread
An instance of Thread which is rooted. See the Thread type for documentation on interacting with the type.

Thread
Representation of the virtual machine

Enums

Error
Error type wrapping all possible errors that can be generated from gluon

Constants

PRELUDE

Functions

filename_to_module

new_vm
Creates a new virtual machine with support for importing other modules and with all primitives loaded.

Type Definitions

Result
Type alias for results returned by gluon