Module gluon::import

Implementation of the import macro.

Structs

CheckImporter

DefaultImporter

Import
Macro which rewrites occurances of import "filename" to a load of that file if it is not already loaded and then a global access to the loaded module

Enums

Error
Error type for the import macro

Traits

Importer