Struct gluon::import::Import

pub struct Import<I = DefaultImporter> {
    pub paths: RwLock<Vec<PathBuf>>,
    pub importer: I,
}

Macro which rewrites occurances of import "filename" to a load of that file if it is not already loaded and then a global access to the loaded module

Fields
paths: RwLock<Vec<PathBuf>> importer: I
Methods

impl<I> Import<I>

fn new(importer: I) -> Import<I>

Creates a new import macro

fn add_path<P: Into<PathBuf>>(&self, path: P)

Adds a path to the list of paths which the importer uses to find files